Fix Criteria Bugs
Fix Inheritance
Fix nullPointer object for delegates
