+++ Back end/Polaris/Elastic/Filtering/Criteria/NestCriteria.cs
@@ -1,4 +1,11 @@
+using System;
 using Nest;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Reflection;
+using System.Collections.Generic;
+using Elastic.Filtering.Attributes;
+
 
 namespace Elastic.Filtering.Criteria
 {
@@ -16,5 +23,37 @@ namespace Elastic.Filtering.Criteria
         protected string Value { get; set; }
 
         public abstract QueryContainer Interpret();
+
+        public static Dictionary<string, Func<TNestCriteria, string, string, QueryContainer>> GetRegistry<TNestCriteria>()
+        where TNestCriteria : NestCriteria
+        {
+            var registry = new Dictionary<string, Func<TNestCriteria, string, string, QueryContainer>>();
+            var methods = typeof(TNestCriteria)
+                .GetMethods(BindingFlags.Public | BindingFlags.Static)
+                .Where(m => m.GetCustomAttributes(typeof(NestOperator), false).Length > 0);
+            foreach(var method in methods){
+                var pair = BuildMethodDelegate<TNestCriteria>(method);
+                registry[pair.Key] = pair.Value;
+            }
+
+            return registry;
+        }
+
+        private static KeyValuePair<string, Func<TNestCriteria, string, string, QueryContainer>> BuildMethodDelegate<TNestCriteria>(MethodInfo method)
+        where TNestCriteria : NestCriteria
+        {
+            var objectInput = Expression.Parameter(typeof(TNestCriteria), "criteria");
+            var fieldInput = Expression.Parameter(typeof(string), "field");
+            var valueInput = Expression.Parameter(typeof(string), "value");
+            var lambdaExpression = Expression.Lambda<Func<TNestCriteria, string, string, QueryContainer>>(
+                Expression.Call(null, method, fieldInput, valueInput), objectInput, fieldInput, valueInput)
+                .Compile();
+
+            return new KeyValuePair<string, Func<TNestCriteria, string, string, QueryContainer>>(
+                ((NestOperator)method.GetCustomAttribute(typeof(NestOperator), false)).Abbrv,
+                lambdaExpression
+            );
+        }
+
     }
 }
